refactor: google sign in/out buttons
- Updated register page max width
- Re-organized sign in/out buttons to components/auth/
- Formatted code and lint free
